introduce properties to vectors
