+++ napari/layers/points/_tests/test_points_utils.py
@@ -1,10 +1,8 @@
 import numpy as np
 import pandas as pd
 
-from napari.layers.points._points_utils import (
-    dataframe_to_properties,
-    guess_continuous,
-)
+from napari.layers.points._points_utils import guess_continuous
+from napari.layers.utils.layer_utils import dataframe_to_properties
 
 
 def test_dataframe_to_properties():
